pyproject.toml
[project]
name = "creatorai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.60",
    "fastapi>=0.135.1",
    "flask>=3.1.3",
    "joblib>=1.5.3",
    "ntscraper>=0.4.0",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "snscrape>=0.7.0.20230622",
    "tweepy>=4.16.0",
    "twikit>=2.3.3",
    "uvicorn>=0.41.0",
    "xgboost>=3.2.0",
]
